+++ src/TokenType.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Majored\PhpBbbApiWrapper;
+
+/** Holds declarations for different API token types. */
+class TokenType {
+    /** @var string A string value representing the Private token type. */
+    public const PRIVATE = "Private";
+
+    /** @var string A string value representing the Shared token type. */
+    public const SHARED = "Shared";
+}
